Add simple support for rspec files with hanami
I did this because the container architecture of hanami makes it so that there
can be many applications under `apps`. This checks to see if we are in
hanami's container architecture before doing the simple rails check. The
function is designed to take the path given and not a modified path so that
useful information about the path can be used.

I am not tackling the hanami lib concerns with these changes.
This does not address opening hanami asset alternates like javascript.