Correct Rails RSpec and Support Hanami Libs
Why you made the change:

I corrected the Rails RSpec integration to match that of the Ruby Gem
support because that is what is officially supported by the rspec-rails
gem. I also added Hanami Lib support so that we could alternate between
hanami libs and spec files.

Looked at the use case of a hanamai app implementation file mapping to a
spec file and back, however it is non trivial since there isn't enough
information in the spec path alone to make its way back to the app path.
It would have to do some more involved searching of the app(s) path
contents to determine things.